fixing the bucket organizer, unit tests adjustments and start the algorithm to choose the package to rturn

+++ src/main/java/com/mobiquityinc/model/Item.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class Item {
     private int index;
     private float weight;
-    private float price;
+    private int price;
 }
